backend/requirements.txt
psycopg2-binary
gunicorn==23.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1
requests==2.32.5
annotated-types==0.7.0
asgiref==3.11.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl#sha256=7d6ad14cd5078e53147bfbf70fb9d433c6a3865b695fda2657140bbc59a27e29
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
murmurhash==1.0.15
nltk==3.8.1
numpy==2.4.2
packaging==26.0
pandas==2.2.3
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2026.1.15
requests==2.32.5
scikit-learn==1.5.2
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
